requirements.txt
numpy==1.18.5
PyYAML==5.3.1
torch
torchvision
easydict==1.9